refactor initial code
