+++ src/main/java/ru/skypro/homework/SimpleAvitoApplication.java
@@ -1,12 +1,14 @@
 package ru.skypro.homework;
 
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+@OpenAPIDefinition
 @SpringBootApplication
-public class HomeworkApplication {
+public class SimpleAvitoApplication {
 
   public static void main(String[] args) {
-    SpringApplication.run(HomeworkApplication.class, args);
+    SpringApplication.run(SimpleAvitoApplication.class, args);
   }
 }
